+++ src/apps/scc.rs
@@ -54,7 +54,7 @@ impl App for Scc {
         yard.load_app(self.name(), version, self.executable_filename(platform))
     }
 
-    fn versions(&self, amount: u8, output: &dyn Output) -> Result<Vec<String>> {
+    fn versions(&self, amount: usize, output: &dyn Output) -> Result<Vec<String>> {
         github_releases::versions(ORG, REPO, amount, output)
     }
 }
